LangChain/KrishNaik/requirements.txt
#langchain_openai
langchain
langchain_core
langchain_community
python-dotenv
streamlit
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
#chromadb
faiss-cpu
langchain-groq
groq
cassio
beautifulsoup4
wikipedia
arxiv
langchainhub
sentence_transformers
PyPDF2
langchain-objectbox